Refactor repository layout.
