+++ build.gradle
@@ -1,0 +1,5 @@
+
+allprojects {
+    group = 'com.ibm.cicsdev.eventprocessing'
+    version = '0.0.1-SNAPSHOT'
+}
