+++ cics-eventprocessing-java-bundle/src/main/bundleParts/ACCOUNTADAPTER.epadapter
