+++ cics-eventprocessing-java-bundle/src/main/bundleParts/EVNT.transaction
