+++ cics-eventprocessing-java-webapp/build.gradle
@@ -1,0 +1,29 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+plugins {
+    id 'war'
+}
+
+repositories {
+    mavenCentral()
+}
+
+java
+{
+    toolchain
+    {
+        languageVersion = JavaLanguageVersion.of(java_version)
+        vendor = JvmVendorSpec.IBM
+        implementation = JvmImplementation.J9
+    }
+}
+
+dependencies {
+    providedCompile 'com.ibm.cics:com.ibm.cics.server:1.700.1-5.5-PH25409'
+    providedCompile 'com.ibm.cics:com.ibm.cics.server.invocation.annotations:5.5-PH25409'
+    providedCompile 'javax.servlet:javax.servlet-api:3.1.0'
+}
+
+description = 'cics-eventprocessing-java - Webapp'
