--- cics-eventprocessing-java/build.gradle
@@ -1,5 +1,0 @@
-
-allprojects {
-    group = 'com.ibm.cicsdev.eventprocessing'
-    version = '0.0.1-SNAPSHOT'
-}
