--- cics-eventprocessing-java/cics-eventprocessing-java-bundle/build.gradle
@@ -1,19 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-plugins {
-    id 'com.ibm.cics.bundle' version '1.0.3'
-}
-
-cicsBundle {
-    build {
-        defaultJVMServer = 'DFHWLP'
-    }
-}
-
-dependencies {
-    //implementation 'com.ibm.cicsdev.eventprocessing:cics-eventprocessing-java-code:0.0.1-SNAPSHOT'
-    cicsBundlePart project(path:':cics-eventprocessing-java-webapp', configuration: 'archives')
-}
-
-description = 'cics-eventprocessing-java - CICS bundle'
