--- cics-eventprocessing-java/cics-eventprocessing-java-webapp/src/main/webapp/WEB-INF/web.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee/" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_1.xsd"
-	version="3.1">
-    <display-name>cics-eventprocessing-java</display-name>
-</web-app>
