+++ pom.xml
@@ -1,0 +1,53 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>com.ibm.cicsdev.eventprocessing</groupId>
+	<artifactId>cics-eventprocessing-java</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
+	<packaging>pom</packaging>
+
+	<name>cics-eventprocessing-java</name>
+
+  <properties>
+	<java.version>1.8</java.version>
+	
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <maven.compiler.target>${java.version}</maven.compiler.target>
+    <maven.compiler.source>${java.version}</maven.compiler.source>
+  </properties>
+  
+  <build>
+      <pluginManagement>
+      <plugins>
+        <plugin>
+          <groupId>com.ibm.cics</groupId>
+          <artifactId>cics-bundle-maven-plugin</artifactId>
+          <version>1.0.3</version>
+        </plugin>
+      </plugins>
+    </pluginManagement>
+  </build>
+  
+  
+
+	<!-- CICS TS V5.5 BOM (as of May 2020) -->
+	<dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>com.ibm.cics</groupId>
+				<artifactId>com.ibm.cics.ts.bom</artifactId>
+				<version>5.5-20200519131930-PH25409</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+		</dependencies>
+	</dependencyManagement>
+  
+  <url>https://github.com/cicsdev/cics-eventprocessing-java</url>
+  <modules>
+  	<module>cics-eventprocessing-java-webapp</module>
+  	<module>cics-eventprocessing-java-bundle</module>
+  </modules>
+</project>
