+++ settings.gradle
@@ -1,0 +1,4 @@
+
+rootProject.name = 'cics-eventprocessing-java'
+include(':cics-eventprocessing-java-webapp')
+include(':cics-eventprocessing-java-bundle')
